refactor(cache): Extract cache primitives from store
